# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgriLog v2 API",
    "description": "API cho ứng dụng nông hộ AgriLog v2 (mobile) và trang quản trị (web-admin).\n\n**Xác thực:** đăng nhập ở `POST /auth/login`, rồi gửi `Authorization: Bearer <token>` cho mọi endpoint khác.\n\n**Đồng bộ:** `GET /sync` kéo thay đổi, `POST /sync` đẩy thay đổi. Xung đột giải quyết theo *last write wins* dựa trên `updated_at`, và mọi bản ghi bị giữ lại đều trả về trong `conflicts` kèm bản của máy chủ.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/warehouse/out": {
      "get": {
        "tags": [
          "warehouse"
        ],
        "summary": "List Warehouse Out",
        "operationId": "list_warehouse_out_warehouse_out_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "owner_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WarehouseOutOut"
                  },
                  "title": "Response List Warehouse Out Warehouse Out Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "warehouse"
        ],
        "summary": "Create Warehouse Out",
        "operationId": "create_warehouse_out_warehouse_out_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseOutCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseOutOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/warehouse/out/{row_id}": {
      "patch": {
        "tags": [
          "warehouse"
        ],
        "summary": "Update Warehouse Out",
        "operationId": "update_warehouse_out_warehouse_out__row_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "row_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Row Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseOutUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseOutOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "warehouse"
        ],
        "summary": "Delete Warehouse Out",
        "operationId": "delete_warehouse_out_warehouse_out__row_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "row_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Row Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/media": {
      "post": {
        "tags": [
          "media"
        ],
        "summary": "Upload Media",
        "description": "Stores one photo or video. A repeat with the same `id` returns 200 and\nthe file already stored, so a phone can retry after a dropped response.",
        "operationId": "upload_media_media_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_media_media_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/media/token": {
      "post": {
        "tags": [
          "media"
        ],
        "summary": "Media Token",
        "description": "A token for `?t=` on media URLs. It opens /media only — never the API.",
        "operationId": "media_token_media_token_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaTokenOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/media/{media_id}": {
      "get": {
        "tags": [
          "media"
        ],
        "summary": "Get Media",
        "operationId": "get_media_media__media_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Media Id"
            }
          },
          {
            "name": "t",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Token từ POST /media/token",
              "title": "T"
            },
            "description": "Token từ POST /media/token"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "media"
        ],
        "summary": "Delete Media",
        "operationId": "delete_media_media__media_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Media Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_upload_media_media_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id",
            "description": "Mã do điện thoại đặt; gửi lại cùng mã là an toàn"
          },
          "public": {
            "type": "boolean",
            "title": "Public",
            "description": "Ảnh công khai (hướng dẫn chăm sóc) — chỉ quản trị viên",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_media_media_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MediaOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "is_public": {
            "type": "boolean",
            "title": "Is Public"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "created_at": {
            "type": "integer",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "content_type",
          "size_bytes",
          "is_public",
          "url",
          "created_at"
        ],
        "title": "MediaOut"
      },
...
      "MediaTokenOut": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "expires_at": {
            "type": "integer",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "token",
          "expires_at"
        ],
        "title": "MediaTokenOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WarehouseOutCreate": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "fertilizer_id": {
            "type": "string",
            "maxLength": 64,
            "title": "Fertilizer Id"
          },
          "fertilizer_name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Fertilizer Name"
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "quantity_kg": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Quantity Kg"
          },
          "occurred_at": {
            "type": "integer",
            "title": "Occurred At"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "plot_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Plot Id"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          }
        },
        "type": "object",
        "required": [
          "fertilizer_id",
          "fertilizer_name",
          "quantity_kg",
          "occurred_at"
        ],
        "title": "WarehouseOutCreate"
      },
      "WarehouseOutOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          },
          "created_at": {
            "type": "integer",
            "title": "Created At"
          },
          "updated_at": {
            "type": "integer",
            "title": "Updated At"
          },
          "fertilizer_id": {
            "type": "string",
            "title": "Fertilizer Id"
          },
          "fertilizer_name": {
            "type": "string",
            "title": "Fertilizer Name"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "quantity_kg": {
            "type": "number",
            "title": "Quantity Kg"
          },
          "unit_price": {
            "type": "number",
            "title": "Unit Price"
          },
          "total_cost": {
            "type": "number",
            "title": "Total Cost"
          },
          "occurred_at": {
            "type": "integer",
            "title": "Occurred At"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "plot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plot Id"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "owner_id",
          "updated_by",
          "created_at",
          "updated_at",
          "fertilizer_id",
          "fertilizer_name",
          "category",
          "quantity_kg",
          "unit_price",
          "total_cost",
          "occurred_at",
          "note",
          "plot_id",
          "plan_id"
        ],
        "title": "WarehouseOutOut"
      },
      "WarehouseOutUpdate": {
        "properties": {
          "fertilizer_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Fertilizer Id"
          },
          "fertilizer_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Fertilizer Name"
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "quantity_kg": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity Kg"
          },
          "unit_price": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Price"
          },
          "occurred_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "plot_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Plot Id"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          }
        },
        "type": "object",
        "title": "WarehouseOutUpdate"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}